--- a/app/models/mixins/aggregation_mixin/methods.rb
+++ b/app/models/mixins/aggregation_mixin/methods.rb
@@ -39,11 +39,8 @@ module AggregationMixin
     end
 
     def aggregate_hardware(from, field, targets = nil)
-      from      = from.to_s.singularize
-      select    = field == :aggregate_cpu_speed ? "cpu_total_cores, cpu_speed" : field
-      targets ||= send("all_#{from.pluralize}")
-      hdws      = Hardware.where(from.singularize => targets).select(select)
-      hdws.inject(0) { |t, hdw| t + hdw.send(field).to_i }
+      targets ||= send("all_#{from.to_s.pluralize}")
+      Hardware.where(from.to_s.singularize => targets).sum(field)
     end
 
     def lans