@kbrock 2016-12-03
aah, of note, this brakeman error is not new. there is currently an entry for this in `brakeman.ignore` - it simply has changed that method enough that we need to tweak the ignore file