@kbrock 2017-04-24
this is 2 PRs.

1. use inner selects for hardware ids
2. use `sum` instead of bringing back all hardware records.

May make sense to pull out the first change.

May make sense to introduce using `vm_hardwares` and `host_hardwares` concept and remove much of this logic.