@kbrock 2018-02-01
Added tests for each of these columns.
They are in a separate PR - figure we'll get those in regardless of whether we want to go this route.