@Ladas 2018-02-01
@kbrock right, a .where condition might appear in the future (doing active/archived by scopes), so if we have specs that is good.

To have working on every query, it should be enough to add 
```
sum("#{table_name}.#{column_name}") # and maybe the same for select
```

Or what was the exact failure you've been seeing with where/unique/etc.?