@NickLaMuro 2019-01-16
@kbrock Is this something I should still review?  Do you still want to get it in?  There is a `[WIP]` label on it, so unsure where things with this are at (also, I guess it is 2 years old at this point...).