@kbrock 2019-10-17
Great addition from @Ladas 

We are close to having `virtual_aggregate` working with `has_many :through`
So we may be able to use the associations directly rather than all this monkey business here.