@kbrock 2019-10-21
comments:
1. Merged @Ladas suggestion here - got rid of a brakeman warning 👮 and a line 🔥 - very cool.
2. this is getting very close to a one liner - which is nice
3. Once we merge the `virtual_total` with `has _many :through`, then we may be able to drop more in this file.