@kbrock 2019-11-18
removed an unnecessary comment change.