@kbrock 2020-05-14
fixed this.
Drew U got in most of these changes.

I'm still holding out until we know that the sum will not overflow an integer