@kbrock 2020-05-22
@d-m-u has taken this and run with it. I opt for her solution